vertex-ai-matching-engine-1/requirements.txt
tensorflow==2.12.0
google-cloud-storage==2.8.0
google-cloud-aiplatform==1.24.1